+++ src/main/java/com/hallym/rehab/domain/room/repository/AudioRepository.java
@@ -1,7 +1,7 @@
 package com.hallym.rehab.domain.room.repository;
 
-import com.hallym.rehab.domain.room.domain.Audio;
-import com.hallym.rehab.domain.room.domain.Room;
+import com.hallym.rehab.domain.room.entity.Audio;
+import com.hallym.rehab.domain.room.entity.Room;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
